// server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  AGENT: 'AGENT',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Purpose = {
  SALE: 'SALE',
  RENT: 'RENT'
} as const

export type Purpose = (typeof Purpose)[keyof typeof Purpose]


export const PropertyType = {
  APARTMENT: 'APARTMENT',
  HOUSE: 'HOUSE',
  DUPLEX: 'DUPLEX',
  LAND: 'LAND',
  COMMERCIAL: 'COMMERCIAL'
} as const

export type PropertyType = (typeof PropertyType)[keyof typeof PropertyType]


export const ListingStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type ListingStatus = (typeof ListingStatus)[keyof typeof ListingStatus]


export const Availability = {
  AVAILABLE: 'AVAILABLE',
  SOLD: 'SOLD',
  RENTED: 'RENTED'
} as const

export type Availability = (typeof Availability)[keyof typeof Availability]


export const InquiryStatus = {
  PENDING: 'PENDING',
  CONTACTED: 'CONTACTED',
  CLOSED: 'CLOSED'
} as const

export type InquiryStatus = (typeof InquiryStatus)[keyof typeof InquiryStatus]